Update scrape_jobs
Updated scrape jobs to clean and extract data from job postings.

This function looks for jobs which match on the job_title and company_name, calculates the similarity score and does not add jobs which have a similarity score above 0.99 with a job which is already in the database. 

This function also extracts the salary and skills from the description.